Add option to auto-resubscribe (enabled by default)
If a user has granted permission for notifications, but has no Kumulos subscription
data stored locally, we subscribe the user for push.

Any existing push subscription is unsubscribed prior to resubscribing.

When auto-resubscription is disabled, prompts will be shown as if the user is not
currently subscribed. Using the prompt will re-subscribe the user without asking
for native permission.