Render UI through a portal to prevent conflicts with existing DOM content
In certain cases, rendering directly into document.body could cause existing DOM
nodes to be replaced with a prompt widget.

By rendering into a dedicated UI root, and then using a portal to move the children
to the body, we achieve the same result without allowing the existing DOM to become
mutated.