Add ability to configure the SDK from the worker (#13)
This assists with migration cases where the existing provider's service worker is
transitioned to ours. In this case, if the user does not visit the page for auto-resubscribe,
the config entry is not generated and our worker fails to load a Context.

This adds a check for a `self.KUMULOS_INIT` object which will be used to persist the minimal
config necessary to build a Context.

In this way, the worker should be able to 'wake up' to handle pushes from Kumulos during migration.