Fix prompt trigger property matching edge case
When a trigger was set up to match '/' using propIn, the trigger was considered to match '/foo' as well, because the
generated regex was new RegExp('\/', 'g'), which matches the slash in any position.

The resolution is to make the regex explicitly test start & end of pattern using the ^ and $ respectively:

new RegExp('^\/$', 'g')

As we support wildcards, we can still achieve prefix & suffix matching.