Include installId in detailsHash components
Ensure that a INSTALL_TRACKED event will be emitted if the installId changes.

This can happen if (for example) the localStorage is cleared. On next page load, the original visitor ID will be regenerated and install ID will change.

Because we try to limit call homes unless necessary, the hash is compared, but if localStorage is cleared and indexedDB is not, the details hash won't change and no INSTALL_TRACKED event will be created.

By including the install ID in the hash it makes sure the old hash won't match after the ID changes.