Remove UndefinedStateUIContext and some dead code for channel subscription
In Optimove, we don't support the Kumulos 'channels' feature.

The original UIContext was used primarily to facilitate the channels use cases.

The UIContext is consumed in the Dialog for alert & banner prompt types.

If the UIContext value is undefined, we don't draw any dialogs.

The solution is to remove the channel code that is dependent on the UIContext and remove UIContext provider.

The channel code is not used because in Optimove there is no way to define this config / it is not a supported feature.